1) Implemented ClientDao.getClient() and FoodDao.getFood() in sql 2) Replaced double type in Food POJO on BigDecimal

+++ src/main/java/com/belenot/eatfood/domain/Food.java
@@ -1,28 +1,30 @@
 package com.belenot.eatfood.domain;
 
+import java.math.BigDecimal;
+
 public class Food {
-    private long id;
+    private int id;
     private Client client;
     private String name;
     //Use BigDecimal
-    private double calories;
-    private double protein;
-    private double carbohydrate;
-    private double fat;
+    private BigDecimal calories;
+    private BigDecimal protein;
+    private BigDecimal carbohydrate;
+    private BigDecimal fat;
     
-    public void setId(long id) { this.id = id; }
+    public void setId(int id) { this.id = id; }
     public void setClient(Client client) { this.client = client; }
     public void setName(String name) { this.name = name; }
-    public void setCalories(double calories) { this.calories = calories; }
-    public void setProtein(double protein) { this.protein = protein; }
-    public void setCarbohydrate(double carbohydrate) { this.carbohydrate = carbohydrate; }
-    public void setFat(double fat) { this.fat = fat; }
+    public void setCalories(BigDecimal calories) { this.calories = calories; }
+    public void setProtein(BigDecimal protein) { this.protein = protein; }
+    public void setCarbohydrate(BigDecimal carbohydrate) { this.carbohydrate = carbohydrate; }
+    public void setFat(BigDecimal fat) { this.fat = fat; }
 
-    public long getId() { return id; }
+    public int getId() { return id; }
     public Client getClient() { return client; }
     public String getName() { return name; }
-    public double getCalories() { return calories; }
-    public double getProtein() { return protein; }
-    public double getCarbohydrate() { return carbohydrate; }
-    public double getFat() { return fat; }
+    public BigDecimal getCalories() { return calories; }
+    public BigDecimal getProtein() { return protein; }
+    public BigDecimal getCarbohydrate() { return carbohydrate; }
+    public BigDecimal getFat() { return fat; }
 }
